Update: Migrate to ES6 syntax (fixes #6)
